Added enrichment compression
Added enrichment compression that checks similarity between two sets of terms. There are still some options to be worked out. One of which is how to sort the array initially. The MAGINE methods allows one to do this on the fly due to the class structure. It is harder in R, so maybe @sgosline can provide some insights how to make it general? Also, in MAGINE one can find terms that were removed in a post analysis step. I think it could be useful here as well. I am debating writing a reticulate wrapper for the magine enrichment, which would bring all the tools in. How to do that as a package, still unsure.